standalone/experiments/multi-agent/multi-mcp-agent-system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
strands-agents
mcp
python-dotenv
pydantic
asyncio
concurrent.futures